+++ CHANGELOG.md
@@ -57,11 +57,17 @@
 - [#4515](https://github.com/ChainSafe/forest/pull/4515) Add support for the
   `Filecoin.StateLookupRobustAddress` RPC method.
 
+- [#4496](https://github.com/ChainSafe/forest/pull/4496) Add support for the
+  `Filecoin.EthEstimateGas` RPC method.
+
 - [#4558](https://github.com/ChainSafe/forest/pull/4558) Add support for the
   `Filecoin.StateVerifiedRegistryRootKey` RPC method.
 
 - [#4474](https://github.com/ChainSafe/forest/pull/4474) Add new subcommand
   `forest-cli healthcheck ready`.
+
+- [#4569](https://github.com/ChainSafe/forest/pull/4569) Add support for the
+  `Filecoin.NetFindPeer` RPC method.
 
 - [#4565](https://github.com/ChainSafe/forest/pull/4565) Add support for the
   `Filecoin.StateGetRandomnessDigestFromBeacon` RPC method.